+++ examples/agrume-cli-example/src/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line filename-rules/not-match
-import './src/app.tsx'
+import './app.tsx'