+++ packages/core/src/create-route/create-route.ts
@@ -1,7 +1,7 @@
+import { state, utils } from '@agrume/internals'
 import type { AnyRoute, Client, CreateRoute, RequestOptions, RouteOptions, RouteReturnValue } from '@agrume/types'
 import babelParser from '@babel/parser'
 
-import { state } from '@agrume/internals'
 import { options } from './options'
 import { getRouteName } from './get-route-name'
 import { getRequestOptions } from './get-request-options'
@@ -21,7 +21,25 @@ export function createRoute<
 > {
   const routeName = getRouteName(route, routeOptions)
   const prefix = options.get().prefix
-  const baseUrl = options.get().baseUrl.slice(0, -1) // .slice(0, -1) removes the trailing slash, because prefix already has it
+
+  let host = ''
+
+  const baseUrl = options.get().baseUrl?.slice(0, -1) // .slice(0, -1) removes the trailing slash, because prefix already has it
+  if (baseUrl !== undefined) {
+    host = baseUrl
+  }
+
+  const tunnelType = options.get().tunnel?.type
+  if (tunnelType !== undefined) {
+    const tunnelInfos = utils.getTunnelInfos(tunnelType)
+
+    if (tunnelInfos.type === 'localtunnel') {
+      host = `https://${tunnelInfos.tunnelSubdomain}.${tunnelInfos.tunnelDomain}`
+    }
+    else if (tunnelInfos.type === 'bore') {
+      host = `http://${tunnelInfos.tunnelDomain}:${tunnelInfos.tunnelPort}`
+    }
+  }
 
   if (state.get()?.isRegistering) {
     state.set((state) => {
@@ -30,7 +48,7 @@ export function createRoute<
     })
   }
 
-  const requestOptions = getRequestOptions(`${baseUrl}${prefix}${routeName}`)
+  const requestOptions = getRequestOptions(`${host}${prefix}${routeName}`)
 
   const getClient = routeOptions?.getClient ?? getDefaultClient
   let stringifiedGetClient = getClient.toString()
@@ -88,11 +106,29 @@ function getDefaultClient<R extends AnyRoute>(
             return
           }
 
-          const value = unformattedValue.startsWith('data: ')
-            ? unformattedValue.slice(6)
-            : unformattedValue
+          const unformattedValues = unformattedValue.split('\n\n')
 
-          yield JSON.parse(value)
+          for (const unformattedValue of unformattedValues) {
+            if (unformattedValue === '') {
+              continue
+            }
+
+            const DATA_PREFIX = 'data: '
+            const data = unformattedValue.startsWith(DATA_PREFIX)
+              ? unformattedValue.slice(DATA_PREFIX.length)
+              : unformattedValue
+
+            if (data === 'DONE') {
+              return
+            }
+
+            const RETURN_PREFIX = 'RETURN'
+            if (data.startsWith(RETURN_PREFIX)) {
+              return JSON.parse(data.slice(RETURN_PREFIX.length))
+            }
+
+            yield JSON.parse(data)
+          }
         }
       }