+++ packages/internals/src/utils/get-tunnel-infos.ts
@@ -1,0 +1,46 @@
+import { machineIdSync } from 'node-machine-id'
+import { checksum } from './checksum'
+
+type TunnelInfos =
+  | { tunnelDomain: 'bore.pub', tunnelPort: number, type: 'bore' }
+  | { tunnelDomain: 'loca.lt', tunnelSubdomain: string, type: 'localtunnel' }
+
+/**
+ * Get the tunnel infos for the given tunnel type.
+ * @param {string} [tunnelType] The tunnel type.
+ * @returns {TunnelInfos} The tunnel URL.
+ */
+export function getTunnelInfos(tunnelType: TunnelInfos['type']): TunnelInfos {
+  switch (tunnelType) {
+  case 'localtunnel': {
+    const machineId = machineIdSync()
+    return {
+      tunnelDomain: 'loca.lt',
+      tunnelSubdomain: checksum(machineId),
+      type: 'localtunnel',
+    }
+  }
+  case 'bore': {
+    const machineId = machineIdSync()
+
+    const removeScientificNotation = (number: string) =>
+      number.split('e')[0]!.replace('.', '').slice(-2)
+
+    const n = Number(
+      removeScientificNotation(Number.parseInt(machineId, 16).toString()),
+    )
+
+    const determinedPort = removeScientificNotation(
+      Number.parseInt(machineId, n % 32).toString(),
+    )
+      ?.repeat(10)
+      .slice(n % 10, n % 10 + 4)
+
+    return {
+      tunnelDomain: 'bore.pub',
+      tunnelPort: Number.parseInt(determinedPort),
+      type: 'bore',
+    }
+  }
+  }
+}