+++ packages/internals/src/utils/utils.ts
@@ -1,5 +1,7 @@
 import { checksum } from './checksum'
+import { getTunnelInfos } from './get-tunnel-infos'
 
 export const utils = {
   checksum,
+  getTunnelInfos,
 }