+++ packages/agrume/CHANGELOG.md
@@ -1,5 +1,63 @@
 # agrume
 
+## 2.0.0
+
+### Major Changes
+
+- 5fb60e7: - Support for stream/iterable route
+  - Agrume CLI
+  - Some fixes
+
+### Minor Changes
+
+- 344d4a0: Add support for tunneling
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [d144095]
+- Updated dependencies [843f053]
+- Updated dependencies [344d4a0]
+- Updated dependencies [034c046]
+- Updated dependencies [4ddd7f6]
+- Updated dependencies [5fb60e7]
+  - @agrume/core@2.0.0
+
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [4ddd7f6]
+  - @agrume/core@2.0.0-beta.5
+
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- Updated dependencies [034c046]
+  - @agrume/core@2.0.0-beta.4
+
+## 2.0.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [843f053]
+  - @agrume/core@2.0.0-beta.3
+
+## 2.0.0-beta.2
+
+### Minor Changes
+
+- 344d4a0: Add support for tunneling
+
+### Patch Changes
+
+- Updated dependencies [344d4a0]
+  - @agrume/core@2.0.0-beta.2
+
 ## 2.0.0-beta.1
 
 ### Patch Changes