+++ packages/babel-preset-agrume/CHANGELOG.md
@@ -1,5 +1,86 @@
 # babel-preset-agrume
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies [cbd08ac]
+  - babel-plugin-agrume@2.1.2
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [3c93814]
+  - babel-plugin-agrume@2.1.1
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [f7ab7dc]
+  - babel-plugin-agrume@2.1.0
+
+## 2.0.0
+
+### Major Changes
+
+- a4a3630: Introduce babel-preset-agrume and automatically find entry file when using Agrume CLI.
+
+### Minor Changes
+
+- 9d503c1: Fix support for Flow with JSX syntax and .js extension
+- 344d4a0: Add support for tunneling
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [344d4a0]
+- Updated dependencies [034c046]
+- Updated dependencies [4ddd7f6]
+- Updated dependencies [5fb60e7]
+  - babel-plugin-agrume@2.0.0
+
+## 2.0.0-beta.6
+
+### Patch Changes
+
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [4ddd7f6]
+  - babel-plugin-agrume@2.0.0-beta.5
+
+## 2.0.0-beta.5
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- Updated dependencies [034c046]
+  - babel-plugin-agrume@2.0.0-beta.4
+
+## 2.0.0-beta.4
+
+### Patch Changes
+
+- babel-plugin-agrume@2.0.0-beta.3
+
+## 2.0.0-beta.3
+
+### Minor Changes
+
+- 344d4a0: Add support for tunneling
+
+### Patch Changes
+
+- Updated dependencies [344d4a0]
+  - babel-plugin-agrume@2.0.0-beta.2
+
+## 2.0.0-beta.2
+
+### Minor Changes
+
+- 9d503c1: Fix support for Flow with JSX syntax and .js extension
+
 ## 2.0.0-beta.1
 
 ### Major Changes