+++ packages/plugin/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @agrume/plugin
 
+## 2.0.3
+
+### Patch Changes
+
+- babel-preset-agrume@2.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- babel-preset-agrume@2.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- babel-preset-agrume@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- 5952d93: Introduce cross-framework plugin \`@agrume/plugin\`.
+
+### Minor Changes
+
+- 9d503c1: Fix support for Flow with JSX syntax and .js extension
+- 24e08ae: Add support for Flow.
+- 344d4a0: Add support for tunneling
+- a4a3630: Introduce babel-preset-agrume and automatically find entry file when using Agrume CLI.
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [d144095]
+- Updated dependencies [843f053]
+- Updated dependencies [9d503c1]
+- Updated dependencies [344d4a0]
+- Updated dependencies [034c046]
+- Updated dependencies [4ddd7f6]
+- Updated dependencies [a4a3630]
+- Updated dependencies [5fb60e7]
+  - @agrume/core@2.0.0
+  - @agrume/types@2.0.0
+  - @agrume/internals@2.0.0
+  - babel-preset-agrume@2.0.0
+
+## 2.0.0-beta.9
+
+### Patch Changes
+
+- 4ddd7f6: Show errors of Bore tunnel
+- Updated dependencies [4ddd7f6]
+  - babel-preset-agrume@2.0.0-beta.6
+  - @agrume/core@2.0.0-beta.5
+  - @agrume/internals@2.0.0-beta.5
+
+## 2.0.0-beta.8
+
+### Patch Changes
+
+- 034c046: Use HTTP instead of HTTPS when using Bore
+- Updated dependencies [034c046]
+  - babel-preset-agrume@2.0.0-beta.5
+  - @agrume/core@2.0.0-beta.4
+  - @agrume/internals@2.0.0-beta.4
+
+## 2.0.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [843f053]
+  - @agrume/core@2.0.0-beta.3
+  - @agrume/internals@2.0.0-beta.3
+  - babel-preset-agrume@2.0.0-beta.4
+
+## 2.0.0-beta.6
+
+### Minor Changes
+
+- 344d4a0: Add support for tunneling
+
+### Patch Changes
+
+- Updated dependencies [344d4a0]
+  - babel-preset-agrume@2.0.0-beta.3
+  - @agrume/core@2.0.0-beta.2
+  - @agrume/internals@2.0.0-beta.2
+  - @agrume/types@2.0.0-beta.2
+
+## 2.0.0-beta.5
+
+### Minor Changes
+
+- 9d503c1: Fix support for Flow with JSX syntax and .js extension
+
+### Patch Changes
+
+- Updated dependencies [9d503c1]
+  - babel-preset-agrume@2.0.0-beta.2
+
 ## 2.0.0-beta.4
 
 ### Minor Changes