+++ packages/types/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @agrume/types
 
+## 2.0.0
+
+### Major Changes
+
+- 5fb60e7: - Support for stream/iterable route
+  - Agrume CLI
+  - Some fixes
+
+### Minor Changes
+
+- d144095: Add support for `baseUrl` command.
+- 344d4a0: Add support for tunneling
+
+## 2.0.0-beta.2
+
+### Minor Changes
+
+- 344d4a0: Add support for tunneling
+
 ## 2.0.0-beta.1
 
 ### Minor Changes